requirements.txt
google-adk
python-dotenv
vertexai
llama_index